-introduced the logic for expanding the css shorthands in a fela plugin
